BANCO_REFERENCIAS/backend/requirements.txt
# Backend - Banco de Referências
# Python 3.12+
# Stack: FastAPI + PostgreSQL (Neon) + Google File Search

# Web Framework (FastAPI - 7x mais rápido que Flask)
fastapi==0.115.0
uvicorn[standard]==0.32.0

# Data Validation & Settings
pydantic==2.10.0
pydantic-settings==2.6.0

# Database (PostgreSQL async)
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.13.0
aiosqlite==0.20.0  # SQLite async para testes

# Google Gemini (OBRIGATÓRIO)
google-genai>=1.0.0

# Environment & Config
python-dotenv==1.0.0

# HTTP Client
httpx==0.28.0

# File uploads
python-multipart==0.0.20

# SSL Certificates
certifi>=2024.0.0

# Authentication (Clerk)
clerk-sdk-python==0.1.0

# Testing
pytest==8.3.0
pytest-asyncio==0.23.6
pytest-cov==5.0.0
pytest-mock==3.14.0

# Cache (Redis)
redis==5.0.1
hiredis==2.3.2  # C extension para melhor performance

# Observability (Prometheus)
prometheus-client==0.20.0

